fix openjdk-src bootimage build
By default, the Avian openjdk-src bootimage build won't run to
completion because the OpenJDK jars refer to classes not included in
OpenJDK, leading to NoClassDefFoundErrors when we try to build the
bootimage.  One possible workaround would be to catch such errors and
just replace the code being compiled with code that throws the error
at runtime.  For now, though, we just tell Avian to build the
bootimage generator, classpath.jar, and libavian.a, which are all we
need for building an application.